Refactor types needed for small and medium tests into file that is included in both contexts
